# specs/001-event-core-flow/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Frontend Chat 3rd Anniversary API
  version: 1.0.0
  description: 1차 행사 입장, 답변, 진행자 운영을 위한 서버 계약
servers:
  - url: /
paths:
...
  /api/participants/logout:
    post:
      summary: 참가자 로그아웃
      operationId: logoutParticipant
      security:
        - participantSession: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204': { description: 세션 폐기 완료 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfError' }
...
components:
  securitySchemes:
    participantSession:
      type: apiKey
      in: cookie
      name: __Host-participant_session
    adminSession:
      type: apiKey
      in: cookie
      name: __Host-admin_session
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32 }
  responses:
...
    Unauthorized:
      description: 유효한 세션 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    CsrfError:
      description: 상태 변경 요청 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            field: { type: string }
...